Make test/test_symlink.fish set exit status depending on whether the test passed

+++ test/test_symlink.fish
@@ -4,10 +4,22 @@ mkdir-cd test_symlink
 
 symlink nonexistant anything 2> error.txt
 
-echo error.txt should have: "symlink: `from` argument `...` does not exist"
-echo here it is:
-cat error.txt
+set desired_pattern "symlink: `from` argument '.*' does not exist"
+# echo error.txt should have: "symlink: `from` argument '...' does not exist"
+# echo here it is:
+set output (cat error.txt)
+if not string match -qr $desired_pattern $output
+    echo no match
+    set result 1
+else
+    set result 0
+end
 
-# clean up. TODO have a environment variable to avoid cleaning up for inspection
-rm error.txt
-rmdir-.
+# clean up
+if string-empty $DEBUG
+    echo cleanup
+    rm error.txt
+    rmdir-.
+end
+
+exit $result
